Disable preprompt newline on initial preprompt render
A newline only serves the purpose to separate the prompt from output,
therefore there is no reason to print a newline in front of the
preprompt when rendered for the first time, nor is there a reason for
one when clearing the screen. This commit takes these two states into
consideration.

Fixes #233.