Show full git status information in the rprompt
Unlike the original pure theme which only shows whether the current
git repository is dirty or not, the full and verbose git status
information is shown in the rprompt with an asynchronous manner.

Includes:  added/stashed/modified/untracked/deleted/conflict, etc.

Currently it depends on prezto's git module. The ahead/behind
information (push/pull arrows) is fetched not via prezto but
as in the original pure theme (which involves another git command,
for the sake of the auto-fetch feature).